Conflicts:
	rpcs3/rpcs3.vcxproj.filters

+++ rpcs3/Emu/SysCalls/lv2/SC_TTY.cpp
@@ -11,7 +11,7 @@ int sys_tty_read(u32 ch, u64 buf_addr, u32 len, u64 preadlen_addr)
 
 int sys_tty_write(u32 ch, u64 buf_addr, u32 len, u64 pwritelen_addr)
 {
-	if(ch < 0 || ch > 15 || len <= 0) return CELL_EINVAL;
+	if(ch < 0 || ch > 15 || (s32)len <= 0) return CELL_EINVAL;
 	if(!Memory.IsGoodAddr(buf_addr)) return CELL_EFAULT;
 	
 	Emu.GetDbgCon().Write(ch, Memory.ReadString(buf_addr, len));
